Context agents
For discussion: this is a first draft to try to clarify context agents.  Please comment!  I'll do a picture next to show everything that can have a context, and add that to the PR - as part of that, I need to figure out how the new transfers might fit with context agent.