First draft UML for an actual release
Maybe 0.1? :)  

This is to request feedback on this UML model of what we have either decided or have discussed a lot or just seems like it can be here at this point.  There are some things that will look different in a RDF based model - vowl coming after feedback on this.  Like UML model requires properties to be part of a class, so some of these will look a bit different. 

I have not been totally diligent on looking for all the properties of some of the entities, will do some scanning of issues, but welcome anyone to notice missing ones.