Removing vf:notApplicable and vf:PairsWith.
vf:PairsWith is an odd one because it's a subset of vf:Action *but* with
the caveat that it has vf:notApplicable, making it a subset+1. Modeling
this in a type system becomes fairly strange, and ultimately looks like
treating vf:notApplicable like `null` (or None) in rust. Given this, it
might make sense to make this `null` value implicit by *not specifying
the value for Action enum values that would have returned notApplicable.

If these properties can return null, then vf:notApplicable becomes
redundant, and by extension so does vf:PairsWith (because its only
purpose is to exist as a union between N/A and Action).